inet: Miscellaneous fixes around IPv6, loopback and tcp
I didn't want to do it like this, but it turns out the end result is
very dependent on everything working well.

Add a test for loopback TCP connections. This test revealed bugs in
net/loopback.cpp (which was rewritten to queue packets like a normal-ish
netif for real HW), a reference leak in tcp_handle_packet, bugs in netif
and loopback netif registration (now a loopback v6 route is also
auto-added).

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>