x86: Add x2APIC support
Add x2APIC support, which allows for faster local APIC register accesses
and more than 255 CPUs. Also fix a latent bug in some special cases of
SMP init when the BSP tried to shootdown a TLB on a CPU that was
considered online but was still initializing itself, and was hung on the
vmalloc lock that the BSP was holding (waiting for AP allocating memory
with IRQs off), which resulted in a deadlock. With this,
smp_sync_control got better debugging support.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>