radix: Add marks
Add marks, similar to xarray marks, that can be set on non-NULL entries,
propagating up the tree and allowing for efficient discovery of
particular tagged entries.

The regular iteration code now makes use of this system for both
iterations on tags and normal iteration, through the use of a special
PRESENT tag that is implicitly maintained by radix tree store() functionality.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>