pipe: Cache a page for each pipe
Cache a page for each pipe, to avoid cases where small writes + small
reads could cause pages ping-ponging between the pipe buffers and the
page allocator. This avoids a lot of trips to the page allocator in this
case. This page is only stolen from freed pipe buffers, so the increase
to RSS should be minimal (empty pipes will not hold a cached page).

Tested using will-it-scale/pipe1 and flamegraph(1).

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>