slab: Drop the lock when refilling magazines
Drop the slab cache lock when refilling magazines in
kmem_cache_alloc_refill_mag. This reduces contention in refill, at the
expense of the backend.

The end goal is to always drop both the slab cache lock and preemption
when doing slab allocation, for page allocation in a non-atomic context.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>