stackdepot: Add stackdepot
Add a stackdepot like LLVM and linux, to store stacks in a relatively
compressed format.
We use a similar design to llvm's sanitizer_common_stackdepot and linux's lib/stackdepot.c.
We have capacity for N slabs of 4 pages each. We manually allocate stacks in slabs and link them
in a hash table. To find a specific trace, we index into the hash table using the hash of the
stack trace. To allocate a new stack trace, we first look for it, then if we can't find it we
simply allocate a new one on a slab and link that.

Lookups have a lockless fast-path since we never remove stacks, and
acquire a per-chain bucket lock + a global depot lock (for allocation) if required.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>