ahci: Fix COMMAND_LIST_WRITE setting
AHCI spec says, in "Serial ATA AHCI 1.3.1 Specification" 4.2.2 "Command
List Structure":
	"Write (W): When set, indicates that the direction is a device write (data from system memory to
device). When cleared, indicates that the direction is a device read (data from device to system
memory). If this bit is set and the P bit is set, the HBA may prefetch data in anticipation of
receiving a DMA Setup FIS, a DMA Activate FIS, or PIO Setup FIS, in addition to prefetching
PRDs."

So we clearly had it all wrong and Write is supposed to be set for
writes.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>