rcu: Add a classic RCU implementation
Add a classic RCU implementation based on OLS2001 ("Read-Copy Update"),
Paul McKenney's RCU dissertation and various early RCU articles on lwn
(https://lwn.net/Kernel/Index/#Read-copy-update).

This implementation contains possibly-patented logic and as such is
available according to the terms of the GNU LGPL 2.0 (only!), to which
all the patents are granted.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>