asan: Fix locking when destroying the ASAN quarantine
Previously, we forgot locking the cache when freeing an object from the
quarantine. This fixes it. While we're at it, add asserts to the slab
queues (but disabled by default, even when debugging, as walking through
a linked list N times is horrendously expensive).

Also, in the meanwhile, add GFP_ATOMIC as the default malloc() flag.
This is bad, but it's better than not being able to kick off paged.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>