filemap: Fix writepages race with truncation
Fix a race where truncate would take the page, lock it, set owner to
NULL, remove from the page cache, unlock. Meanwhile the writepages
code already had the page, locks it, looks at it - oops, it has been
truncated, owner is bogus. Detect it.

Note that this is still a problem with other races - namely mapping and
truncate, read/write and truncate. Linux takes the page lock when
mapping, but I don't like that solution, it'll scale horribly.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>