dup2: Fix OOB write on the old fd table
We were erroneously dereferencing ioctx->table, then possibly expanding
the table (creating a new one!), then continuing using the old file
table. This resulted in OOB accesses (because we'd be accessing a file
descriptor index that we had to grow to get there), but not a UAF (due
to sheer magic of this being RCU freed).
Repro: Run /bin/bash and watch it dup2 to a FD high up (255).
Fixes: 9edacb12a ("file: Add RCU implementation of the file descriptor table")

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>