reclaim: Add way to reclaim mapped pages
Add a mapcount counter to struct page, by slashing unsigned long
refcount in 2 (2^32 should work for all refcounts). We maintain this
mapcount when mapping/unmapping, which then reclaim easily uses to check
if it can reclaim this page (or if it's pinned by someone else). This
change required some consolidation of file mapping paths in filemap.cpp,
and of changing all "pmap" implementations. ARM64 was not changed, but
builds.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>