vm_obj: Tolerate page mapcount > 0 after unmapping
So, it turns out we were assuming we assuredly had mapcount = 0 after
unmapping, since we held the lock. This somewhat makes sense, but misses
the edge case of a fork happening (incrementing mapcount) and not being
seen by the unmap code. So, instead of CHECK()ing for mapcount = 0, just
return failure to remove the page. This should activate the page and
guarantee forward progress.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>