vm: Replace amaps with directly storing anon pages in page tables
Instead of maintaing a separate data layer (amaps), store the pages in
the page tables. After mapping a new anon, put the page. This results in
the page's mapcount working like an automatic anon page refcount. While
on it, we also implement an optimization where CoW'ing anon pages with
mapcount = 1 do not get copied, but rather reused and remapped writable.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>