mm: Add swapping
Add swapping of anon pages to disk. Also, while we're at it, perform _a
lot_ of miscellaneous refactoring and moving around of code, including
adding proper GFP flags in a bunch of places, some CHECKs, a bunch of
page fault work. This is not perfect but I'm working on it.

Also fix the UART8250 ACPI and DT builds.
This disables the ARM64 builds in CI because they're broken.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>