net: Rework neighbours
Add a brand new neighbour system, more correct, non-blocking, and takes
full advantage of RCU. This commit does not yet support failure.

While we're at it, we add small fixes on the IPv4 and v6 core code. One
notable fix is the adding of longest matching prefix logic (which is
required for e.g proper IPv4 routing on local networks).

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>