ext2: Keep going on getdirent when we find an unused entry
Stopping on inode = 0 is wrong, because this is simply an ignored entry,
not the end of the list. This repro'd in directories we were simply not
able to delete, because they looked empty from getdirent's PoV, but not
from elsewhere (open and ext2_dir_empty did the right thing). Fix it.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>