x86: Make int3 accessible to userspace
Normally, we set the required CPL to 0 for every interrupt gate. This
makes sense (there is usually no reason to trigger these from userspace
directly using the int instruction, and some of these instructions have
errnos). However, int3 is an exception, where we should be getting a
SIGTRAP from userspace; however, without CPL = 3, the int3 generates a
GPF, which triggers a SIGSEGV.

Fix it.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>