net: Pass iovec_iter around the stack
Instead of getting a normal struct msghdr, create a new kernel_msghdr
that uses a struct iovec_iter instead of struct iovec. This makes it possible
to do zero-copy splice and other sorts of fun tricks with sockets. It also
simplifies and removes quite a bit of handrolled iovec and copy code.

This patch is general cleanup and should not affect behavior.

Fixes: #94
Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>