tcp: Various tcp_input fixes (and SYN-ACK retransmission)
The big change here is SYN-ACK retransmission - which is required. We
also add some preliminary filtering of packets to make sure we don't
spawn a bad socket too early (and subsequently stop the retransmission of
SYN-ACK packets) - this filtering isn't too great nor too correct, but it
is what it is, to be handled properly Later(tm).

Coupled with this is a few error path fixes, and removal of a stray debugb printk.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>