libk/sprintf: Fix snprintf calls with len = 0
When len = 0, we may not write a null terminator (nor any sort of character),
but still need to count to-be-printed characters as normal. The old behavior
did n - 1 as "may use", which ended up underflowing and causing snprintf len = 0
to write all over the place, with no length, causing buffer overflows. This was
especially easy to catch in seq_file code.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>