x86/smp: Set the cpu online mask on boot cpu
As-is, it is possible for the boot cpu to observe an incomplete online cpu
mask really early in boot, around the time where we create new threads. This
causes havoc for us as we try to set CPU affinities for the early threads,
resulting in us accidentally forgetting a CPU.

Instead of letting the AP set the online bit, lets do it from the boot cpu, as
has already been done for riscv64 smp since it was first added.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>