pid: Remove user-controllable integer overflows from lookup
The kernel_api_test SetPgrpTest.HandlesBadPids can trigger integer overflow
and UBSAN warnings from unprivileged userspace due to the way we're calculating
the "max" for maple tree lookups.

To fix it, we must first check if it's not < 0 (if it is, it's not a valid pid
anyway); then, cast it to an unsigned int (we've proven the casting will
not change the value, since > 0) and only then can we + 1.

Fixes: 1a5c11ea3ff6 ("pid: Rework pids and pid management")
Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>