sched: Fix the return value of sched_getaffinity
sched_getaffinity(2) says:
"On  success,  the raw sched_getaffinity() system call returns the number of
bytes placed copied into the mask buffer; this will be the minimum of
cpusetsize and the size (in bytes) of the cpumask_t data type that is used
internally by the kernel to represent the CPU set bit mask."

Thus, fix the return value to return the number of bytes copied. We don't track
the max cpu id we've seen (maybe we should, eventually), so we always transfer
at least the size of a struct cpumask.

This change in return value also allows us to skip the user_memset, which is
done by the libc itself[1][2].

[1] https://github.com/heatd/musl/blob/0784374d561435f7c787a555aeab8ede699ed298/src/sched/affinity.c#L21
[2] https://sourceware.org/git/?p=glibc.git;a=blob;f=sysdeps/unix/sysv/linux/sched_getaffinity.c;h=1c66bebff67ccdc5053f363f030aa53b1b64c3e8;hb=refs/heads/release/2.41/master#l40

Link: https://man7.org/linux/man-pages/man2/sched_setaffinity.2.html
Fixes: 067286cd3d63 ("sched: Add sched_getaffinity and sched_setaffinity")
Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>