x86: Unset the global bit in page tables
On x86, the G bit for page tables is fully unused. For Intel machines,
this is harmless, as the CPU will just ignore it. For AMD machines, the
CPU will actively check it and error out (i.e on the AMD manual, it's
marked as reserved).

Unsetting it fixes booting on AMD cpus.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>