x86: Fix TSC timer behavior when frequency > U32_MAX
On certain cpus (like my AMD Ryzen 9 9900X), the TSC frequency can be
advertised/measured to be larger than UINT32_MAX. Our old logic would
assert on this, and would fall flat on its face.

To avoid this, switch to a new cyc2ns/ns2cyc conversion logic taken from
Linux. Namely, we calculate mult and shift and use those as the main
parameters to do these sorts of conversions.

For now, we're only doing this logic in tsc.cpp, but it really should be
taken to generic clocksource and timer logic.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>