bootmem: Add top-down bootmem allocations
On machines with e.g large amounts of memory or KASAN, it is trivial to
exhaust low mem (DMA32) regions by doing early boot allocations. To
fight this, add a mode to do top-down memory allocations. This should
only be done for callers that are late enough to tolerate 64-bit
addresses.

Arguably, this should be the reverse - all allocations are top-down
unless we say so. However, this would be prone for more breakage,
particularly in early boot architecture code where we're allocating page
tables really early and whatnot. So, we err on the safe side.

Without this, QEMU with 16GB and a KASAN kernel would fail to boot as we
would exhaust all DMA32 reserves before they even reached the page
allocator.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>