tcp: Remove erroneous DCHECK in tcp_destroy_sock
It turns out we can call tcp_destroy_sock with and without the lock. If
we call it with the lock, we can't unref it (without taking a ref of our
own) since the socket can't disappear under us, we need to unlock it
after it.

So remove the DCHECK and document this around the tcp_ destroy_sock
->unref() call.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>