sched: Send a resched on thread_add if needed
Send a resched message to remote CPUs (or set the resched flag on
ourselves) on thread_add if the running thread has a lower priority
than the thread we have just added.

We do this already for thread waking up, although because of an
oversight this was not happening for the first scheduling of a thread.

This patch results in *very large* speed benefits (10x or more) on
workloads that spawn many threads, like a build.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>