mm: Add growing program stacks
Add growing program stack support to the VM subsystem. This includes
logic to grow the stack on page fault, proper reporting of stack areas
in /proc, stack ASLR support and MAP_GROWSDOWN.

The stack now starts at the very top (right under the address space's end),
and ranges down to a certain stack gap (with a gap of 32 pages always
required between the stack's bottom and the highest mmap).

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>