mm: Fix vm_object page removal on reclaim
Before aee336075f59, simply holding the vm_object's lock would mean
that vm_obj_remove_page could simply exclude itself from anyone
trying to grab a page. Now, after that commit, vmo_get stopped
taking the lock. This manifested itself in brokenness, with reclaim
racing with filemap code trying to grab a page from the cache.

To fix that, we introduce the concept of a "frozen page", which is
simply a page with refcount = 0. The code in vmo_get() already
recognized such a concept, with page_try_get() aborting() if refcount =
0. This concept is used in a change in semantics, where
vm_obj_remove_page() itself freezes the page before removing it and
returning (note that freezing is an atomic operation, and may fail).

Fixes: aee336075f59 ("filemap: Remove the locking in vmo_get")
Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>