fs/namei: Do not back out on the last component for depth > 0
When iterating paths, we obviously want to back-out of the last path
segment to get its parent. However, we must not do that for paths when
depth > 0, as we are traversing a symlink and thus a very trivial
example like:

unlink("/bin/a") where /bin is a symlink to /usr/bin, will fail, as:

/ -> /bin -> /usr/bin -> last path segment, back to /usr -> a not found

Fix it by doing the back out in a different part of the code.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>