fs: Delay char/block device lookup to open(2) time
As things were, we did the device lookup in inode_special_init(). This
is wrong and made it so any attempts at looking up such inodes would
fail (e.g stat, etc would not even work themselves).

Instead of that, use ->on_open to delay initialization to open(2) time,
which is the only place where we actually _need_ to do the lookup.

This is possible and correct now that inodes and files are independent.

Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>