fs/dcache: Fix parent list removal on d_destroy
Since the linked commit, we unlink ourselves from d_parent_dir_node iff
we were hashed. This was done by an "if (dentry->d_flags &
DENTRY_FLAG_HASHED)". However, since before it we do the same check, and
unhash ourselves - including the clearing of this flag - this check was
never true.

This (probably) fixes UAFs in dcache code.

Fixes: 99064721f817 ("fs/dcache: Do not null out d_parent on unlink")
Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>