fs/dentry: Fix dcache chain traversal on rename()
rename(), naturally, potentially involves moving one dentry from one
chain to another. This has funky effects on traversal - one can
potentially end up in a different chain than the one it started on.

This is (in theory) ok, because we have the rename_lock seqlock making
us retry if we indeed did not find what we were looking for.

However, since this is list_head traversal, it is expected that we find
the same list_head, in order to finish traversal. If the traversal goes
across chains, it will never find the initial list_head and will thus
keep looping forever and ever (and will also interpret the other chain's
list_head as a proper dentry, which it isn't).

Fix it (in, admittedly, a half-assed way) by doing a ranged check. If we
find any of the hashtable's list_heads, we finish traversal.

Fixes: d5de9db06fe3 ("dentry: RCU-ify dentries and some other bits")
Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>