fs/dcache: Fix dcache parent_dir removal
So, it turns out current semantics are broken.
Fix it by carefully (but, possibly half-assedly) checking if indeed the
list_head is inserted somewhere, in d_destroy. If so, remove it. This
(correctly) decouples the presence in the parent's child list from the
dentry being hashed or not.

While we're at it, we fix dentry lookup failing and make sure we never
retain a dentry that had a failed lookup.

Fixes: 99064721f817 ("fs/dcache: Do not null out d_parent on unlink")
Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>