fs/mount: Fix deadlocks when doing mounts_seq_show()
Since seqlocks got fixed (see fixes tag below), it is possible to
observe a deadlock in mounts_seq_show, namely in seq_d_path_under_root,
where it tries to read_seqlock the mount_lock, and later write_seqlock
when it detects a write.

What is really desired here is read_seqlock_excl, and
read_sequnlock_excl, which provide the semantics wanted here - we're not
writing, but merely want to grab the lock to exclude writers.

Fixes: 6aaafd918c31 ("seqcount: Fix write seqcount's implementation")
Fixes: 62fde32541df ("proc: Add procfs")
Signed-off-by: Pedro Falcato <pedro.falcato@gmail.com>